requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
python-multipart==0.0.6
anthropic==0.8.1
streamlit==1.31.1
PyPDF2==3.0.1
pillow==10.2.0
python-dotenv==1.0.1